JENKINS-42649 - ctrl/cmd + s to open the load/save dialog
